Switched from InstructionsTensor<double> to InstructionsTensor<float> bcuz of nan issues

+++ AI/include/Environment/quantum_circuit_environment.hpp
@@ -155,7 +155,7 @@ public:
    * @return Blob Tensor of 1-axis shape {N×IRS} containing the observation of
    * the current circuit.
    */
-  InstructionsTensor<double> get_observation() const;
+  InstructionsTensor<float> get_observation() const;
 
   /**
    * N = Nr of instructions in the quantum circuit
